Reorganize store module, add example dispatch in selection callback
