--- src/actions/consts.ts
@@ -1,1 +1,0 @@
-export const EDITOR_VISIBLE_CHANGE = 'EDITOR_VISIBLE_CHANGE';
