+++ src/containers/EditorContainer.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { connect } from 'react-redux';
 import Editor from '../components/editor/Editor';
-import {IEditorState} from "../reducers/editor";
+import { IEditorState } from 'store/editor';
 
 interface IEditorContainerProps {
   editor?: IEditorState;
