+++ src/core/bootstrap.ts
@@ -1,4 +1,6 @@
 import { Highlighter, TextSelector } from './index';
+import store from 'store';
+import { textSelectedAction } from 'store/actions';
 
 let handlers;
 
@@ -17,6 +19,5 @@ export function deinitializeCoreHandlers() {
 }
 
 function textSelectorCallback(selection, event) {
-  // TODO logic
-  console.log('selection: ', selection);
+  store.dispatch(textSelectedAction(selection));
 }
