--- src/reducers/index.ts
@@ -1,6 +1,0 @@
-import { combineReducers } from 'redux';
-import editor from './editor';
-
-export default combineReducers({
-  editor,
-});
