+++ src/store/actions.ts
@@ -1,0 +1,2 @@
+export * from './editor/actions';
+export * from './selector/actions';
