+++ src/store/selector/actions.ts
@@ -1,0 +1,7 @@
+export const TEXT_SELECTED = 'TEXT_SELECTED';
+export function textSelectedAction(payload) {
+  return {
+    type: TEXT_SELECTED,
+    payload,
+  };
+}
