+++ src/store/store.ts
@@ -1,8 +1,8 @@
 import { createStore, applyMiddleware } from 'redux';
 import thunk from 'redux-thunk';
-import rootReducer from './reducers';
+import rootReducer from './reducer';
 import promise from 'redux-promise';
-import {createLogger} from 'redux-logger';
+import { createLogger } from 'redux-logger';
 
 const logger = createLogger();
 const store = createStore(
